Implement manage and unmanage methods, update reset
`manage` enables using promiser with deferred objects created by other
means such as `$.ajax()` or manually using `$.Deferred`.

`unmanage` enables removing a deferred object from promiser and
returning it. In case a deferred needs to simply go away, simply
call unmanage to deference it so it falls out of scope.

`reset` has been update to use `unmanage` internally and no longer
enforces the deferred to be not pending. The difference between reset
and unmanage is that `reset` returns the promiser object and unmanage
returns the (now) unmanaged deferred.

Close #4