Implement the watch method
This provides a way for handlers to execute when the deferred is
initially created. For example, an ajax request should not occur, unless
the data will be used on response. The first time a deferred is 'bound'
to via done, then, fail, or always, the watch handlers will be executed
once.

Close #5